+++ modules/store_module.rb
